Removed test installer and just made a DryRun flag

+++ Installer.sh
@@ -2,6 +2,8 @@
 
 set -u
 
+DryRun = false
+
 if [[ $EUID -ne 0 ]]; then
    echo "This script must be run as root" 
    exit 1
@@ -28,7 +30,7 @@ Packages=(
 	virt-manager
 )
 
-secs=10
+countdownSeconds=10
 
 # Boot config file to add IOMMU too (if not there)
 BOOT_CONFIG_FILE="/etc/default/grub"
@@ -43,7 +45,11 @@ get-apt () {
 	echo -e "\n             Getting $1 "
 	echo -e "--------------------------------------------------\n"	
 	
-	apt-get install $1 -y &> /dev/null
+	if [ $DryRun = true ]; then
+		echo "apt-get install $1 -y &> /dev/null"
+	else
+		apt-get install $1 -y &> /dev/null
+	fi
 }
 
 
@@ -52,10 +58,10 @@ echo -e "\n${RED}"
 echo -e "   THIS SCRIPT WILL REBOOT WHEN IT IS FINISHED    "
 
 # Countdown
-while [ $secs -gt 0 ]; do
-	echo -ne "       It will start in $secs seconds.\033[0K\r"
+while [ $countdownSeconds -gt 0 ]; do
+	echo -ne "       It will start in $countdownSeconds seconds.\033[0K\r"
 	sleep 1
-	: $((secs--))
+	: $((countdownSeconds--))
 done
  
 # Update APT before installing
@@ -63,15 +69,23 @@ echo -e "\n${ENDCOLOR}--------------------------------------------------"
 echo -e "                  Updating APT                    "
 echo -e "--------------------------------------------------\n"
 
-apt update
+if [ $DryRun = true ]; then
+	echo "apt update"
+else
+	apt update
+fi
 
 # Loop to call get-apt function
 for i in "${Packages[@]}"; do get-apt "$i"; done
 
 # Check if IOMMU is enabled
 if ! grep -i -q "amd_iommu" "$BOOT_CONFIG_FILE"; then
-    $(sed -i 's/GRUB_CMDLINE_LINUX_DEFAULT="/&amd_iommu=on iommu=pt /' "$BOOT_CONFIG_FILE")
-    grub-update
+	if [ $DryRun = true ]; then
+		echo "Edit $BOOT_CONFIG_FILE and add amd_iommu=on iommu=pt to GRUB_CMDLINE_LINUX_DEFAULT"
+	else
+    	$(sed -i 's/GRUB_CMDLINE_LINUX_DEFAULT="/&amd_iommu=on iommu=pt /' "$BOOT_CONFIG_FILE")
+    	grub-update
+	fi
 else 
     echo "AMD IOMMU All Good!"
 fi
@@ -80,18 +94,37 @@ fi
 if [ -d "/etc/libvirt/hooks" ]; then
 	echo "Libvirt hooks folder found. Leaving Alone."
 else
-	cp -R ./hooks/ /etc/libvirt/
+	if [ $DryRun = true ]; then
+		echo "cp -R ./hooks/ /etc/libvirt/"
+	else
+		cp -R ./hooks/ /etc/libvirt/
+	fi
 fi
 
 # Add user to libvirt group
-usermod -a -G libvirt $(whoami)
+if [ $DryRun = true ]; then
+	echo "usermod -a -G libvirt $(whoami)"
+else
+	usermod -a -G libvirt $(whoami)
+fi
 
 # Check if libvirt service is running
 if [ $(systemctl is-active libvirtd) ]; then
 	echo "Libvirt Service found. Leaving Alone."
 else
-	systemctl start libvirt
-    systemctl enable libvirt
+	if [ $DryRun = true ]; then
+		echo "systemctl start libvirt"
+		echo "systemctl enable libvirt"
+	else
+		systemctl start libvirt
+    	systemctl enable libvirt
+	fi
 fi
 
-#reboot now
+if [ $DryRun = true ]; then
+	echo "Rebooting in 10 seconds"
+else
+	echo "Rebooting in 10 seconds"
+	sleep 10
+	reboot now
+fi
